% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\docType{data}
\name{accident_2013.csv.bz2}
\alias{accident_2013.csv.bz2}
\title{One of three files included in this package}
\description{
One of three files included in this package
}
\references{
\url{https://www.nhtsa.gov/research-data/fatality-analysis-reporting-system-fars}
The dataset contains the following files:
\describe{
 \item{STATE}{}
 \item{ST_CASE}{}
 \item{VE_TOTAL}{}
 \item{VE_FORMS}{}
 \item{PVH_INVL}{}
 \item{PEDS}{}
 \item{PERNOTMVIT}{}
 \item{PERMVIT}{}
 \item{PERSONS}{}
 \item{COUNTY}{}
 \item{CITY}{}
 \item{DAY}{}
 \item{MONTH}{}
 \item{YEAR}{}
 \item{DAY_WEEK}{}
 \item{HOUR}{}
 \item{MINUTE}{}
 \item{NHS}{}
 \item{ROAD_FNC}{}
 \item{ROUTE}{}
 \item{TWAY_ID}{}
 \item{TWAY_ID2}{}
 \item{MILEPT}{}
 \item{LATITUDE}{}
 \item{LONGITUD}{}
 \item{SP_JUR}{}
 \item{HARM_EV}{}
 \item{MAN_COLL}{}
 \item{RELJCT1}{}
 \item{RELJCT2}{}
 \item{TYP_INT}{}
 \item{WRK_ZONE}{}
 \item{REL_ROAD}{}
 \item{LGT_COND}{}
 \item{WEATHER1}{}
 \item{WEATHER2}{}
 \item{WEATHER}{}
 \item{SCH_BUS}{}
 \item{RAIL}{}
 \item{NOT_HOUR}{}
 \item{NOT_MIN}{}
 \item{ARR_HOUR}{}
 \item{ARR_MIN}{}
 \item{HOSP_HR}{}
 \item{HOSP_MN}{}
 \item{CF1}{}
 \item{CF2}{}
 \item{CF3}{}
 \item{FATALS}{}
 \item{DRUNK_DR}{}
 }
}
\author{
Joe Tapper \email{j.tapper@coursera.org}
}
\keyword{data}
