% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{fars_map_state}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{an integer representing the number of the state as determined by the state numbers in the data.
A state number for which there is no data generates an error.}

\item{year}{an integer representing the year associated with the data}
}
\value{
a data frame of with the month and year of each accident summarised by year and month.
}
\description{
fars_map_state : This function takes a state number and year as a parameters
and generates a map of the state showing the location of the accidents for that year.
An invalid STATE number generates an error.
}
\examples{
\dontrun{
fars_map_state()
fars_map_state(10,2014)
}
}
