% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{fars_read}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{The full path to the file containing the fars data.}
}
\value{
a data.frame object
}
\description{
fars_read : This function reads a file into a data.frame object.
}
\examples{
df<-fars_read(make_filename(2013))
head(df)
}
