% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{fars_read_years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{an list of integers representing the years associated with the data}
}
\value{
a data frame of with the month and year of each accident grouped  based on the year.
}
\description{
fars_read_years : This function takes a list of years as a parameter and uses it to build the filename for that year
based on the appropriate naming convetion to generate a dataframe of the years and months of each accident grouped by year.
}
\examples{
fars_read_years(c(2013,2014))
}
