% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{fars_summarize_years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{an list of integers representing the years associated with the data}
}
\value{
a data frame of with the month and year of each accident summarised by year and month.
}
\description{
fars_summarize_years : This function takes a list of years as a parameter and uses it to build the filename for that year
based on the appropriate naming convetion to generate a dataframe of the years and months of each accident summarised by month with each year as a separate variable.
}
\examples{
fars_summarize_years(c(2013,2014))
}
