% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{make_filename.}
\usage{
make_filename(year)
}
\arguments{
\item{year}{an integer representing the year associated with the data}
}
\value{
a string indicating the data file name based on the year.
}
\description{
make_filename : This function takes a year as a parameter and uses it to build the filename for that year
based on the appropriate naming convetion.
}
\examples{
make_filename(2013)
}
